Refactoring of ServiceBus, removes some dublicated code
